26.2: complete BlockEntity sync ValueIO wave + item/registration API fixes
- Migrate all 36 writeSyncNBT/readSyncNBT to ValueOutput/ValueInput
- Base TileThaumcraft: getUpdateTag(HolderLookup.Provider) -> saveWithoutMetadata
- FluidTank -> putChild/readChild; AspectList ValueIO overloads; ItemStack.OPTIONAL_CODEC
- byte[] grids -> putIntArray; contains -> keySet(); getShortOr int cast
- Fix TileBanner/TileSmelter/TileMemory/TileMirror/TileInfusionMatrix/TileWaterJug
- Item API: hurtEnemy void, releaseUsing boolean, appendHoverText/inventoryTick sigs,
  UseAnim -> ItemUseAnimation, getUseDuration(+user)
- ModEntities build(ResourceKey), ModMenuTypes IMenuTypeExtension, MobSpawnType->EntitySpawnReason,
  particle getRandom->random, Registries imports
- Update README + TODO with current status

+++ src/main/java/thaumcraft/common/tiles/crafting/TileVoidSiphon.java
@@ -170,7 +170,7 @@ public class TileVoidSiphon extends TileThaumcraftInventory implements WorldlyCo
     @Override
     public void loadAdditional(ValueInput input) {
         super.loadAdditional(input);
-        progress = input.getShortOr("progress", (short)0);
+        progress = (short) input.getShortOr("progress", (short)0);
     }
     
     // ==================== Inventory ====================
